Moving all state in WPF app to single class. Re-enabling support for Plugins Moving TunnelRelay WPF project into UI folder and renaming it to TunnelRelay.Windows
