+++ TunnelRelay.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26730.8
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TunnelRelay", "TunnelRelay\TunnelRelay.csproj", "{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TunnelRelay.Plugins", "TunnelRelay.Plugins\TunnelRelay.Plugins.csproj", "{5EB32488-8C81-4E8F-ACFF-2C457F5F95F1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TunnelRelay.Core", "TunnelRelay.Core\TunnelRelay.Core.csproj", "{165BB633-EE2D-473C-9D5C-BE92E02FC5CA}"
@@ -21,6 +19,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TunnelRelay.Console.Windows
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TunnelRelay.Diagnostics", "TunnelRelay.Diagnostics\TunnelRelay.Diagnostics.csproj", "{ECA74BCE-20FD-48C9-9642-48492BF913E2}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TunnelRelay.Windows", "UI\TunnelRelay.Windows\TunnelRelay.Windows.csproj", "{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}"
+EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		UI\TunnelRelay.Console.Shared\TunnelRelay.Console.Shared.projitems*{07d399aa-6a69-432b-9852-bc0b52dc4224}*SharedItemsImports = 13
@@ -33,14 +33,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Debug|x86.Build.0 = Debug|Any CPU
-		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Release|x86.ActiveCfg = Release|Any CPU
-		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Release|x86.Build.0 = Release|Any CPU
 		{5EB32488-8C81-4E8F-ACFF-2C457F5F95F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5EB32488-8C81-4E8F-ACFF-2C457F5F95F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5EB32488-8C81-4E8F-ACFF-2C457F5F95F1}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -87,6 +79,14 @@ Global
 		{ECA74BCE-20FD-48C9-9642-48492BF913E2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{ECA74BCE-20FD-48C9-9642-48492BF913E2}.Release|x86.ActiveCfg = Release|Any CPU
 		{ECA74BCE-20FD-48C9-9642-48492BF913E2}.Release|x86.Build.0 = Release|Any CPU
+		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Debug|x86.Build.0 = Debug|Any CPU
+		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Release|x86.ActiveCfg = Release|Any CPU
+		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -95,6 +95,7 @@ Global
 		{07D399AA-6A69-432B-9852-BC0B52DC4224} = {93A42E82-8F83-4AD0-96A5-D0CF6A30483B}
 		{F146DACB-41E4-4BBA-80DD-DB1C5CB8E159} = {93A42E82-8F83-4AD0-96A5-D0CF6A30483B}
 		{A594E3A4-ABC9-49FB-B0CC-604C9FFE670F} = {93A42E82-8F83-4AD0-96A5-D0CF6A30483B}
+		{1D6E50C4-1F9F-4F8A-8916-D3AEABA4A8E0} = {93A42E82-8F83-4AD0-96A5-D0CF6A30483B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9A54A758-13AB-4B23-850F-8730A662F763}
