--- TunnelRelay/App.xaml
@@ -1,10 +1,0 @@
-﻿<Application x:Class="TunnelRelay.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:TunnelRelay"
-             StartupUri="Startup.xaml"
-             x:ClassModifier="internal">
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
