--- TunnelRelay/App.xaml.cs
@@ -1,62 +1,0 @@
-﻿// <copyright file="App.xaml.cs" company="Microsoft">
-// Copyright (c) Microsoft. All rights reserved.
-// </copyright>
-// Licensed under the MIT license.
-// MIT License:
-// Permission is hereby granted, free of charge, to any person obtaining
-// a copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to
-// permit persons to whom the Software is furnished to do so, subject to
-// the following conditions:
-//
-// The above copyright notice and this permission notice shall be
-// included in all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED ""AS IS"", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-namespace TunnelRelay
-{
-    using System.Threading;
-    using System.Windows;
-    using TunnelRelay.Diagnostics;
-    using TunnelRelay.Engine;
-
-    /// <summary>
-    /// Interaction logic for App.xaml
-    /// </summary>
-    internal partial class App : Application
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="App"/> class.
-        /// </summary>
-        public App()
-        {
-            Logger.RegisterLogger(new TextLogger());
-        }
-
-        /// <summary>
-        /// Raises the <see cref="Application.Exit" /> event.
-        /// </summary>
-        /// <param name="e">An <see cref="ExitEventArgs" /> that contains the event data.</param>
-        protected override void OnExit(ExitEventArgs e)
-        {
-            Logger.LogInfo(CallInfo.Site(), "Exiting the application with exit code '{0}'", e.ApplicationExitCode);
-
-            ApplicationData.Instance.SaveSettings();
-            if (TunnelRelay.MainWindow.hybridConnectionManager != null)
-            {
-                TunnelRelay.MainWindow.hybridConnectionManager.CloseAsync(CancellationToken.None).Wait();
-            }
-
-            Logger.Close();
-        }
-    }
-}
