--- TunnelRelay/Engine/HeaderDetails.cs
@@ -1,18 +1,0 @@
-﻿namespace TunnelRelay.Engine
-{
-    /// <summary>
-    /// Header details to store header information for UI.
-    /// </summary>
-    public class HeaderDetails
-    {
-        /// <summary>
-        /// Gets or sets the name of the header.
-        /// </summary>
-        public string HeaderName { get; set; }
-
-        /// <summary>
-        /// Gets or sets the header value.
-        /// </summary>
-        public string HeaderValue { get; set; }
-    }
-}
