--- TunnelRelay/UIPages/PluginManagement.xaml
@@ -1,70 +1,0 @@
-﻿<Window x:Class="TunnelRelay.PluginManagement"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TunnelRelay"
-        mc:Ignorable="d"
-        Title="Plugin Management" Height="700" Width="900">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="10" />
-            <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="10" />
-        </Grid.ColumnDefinitions>
-
-        <Grid.RowDefinitions>
-            <RowDefinition Height="10"/>
-            <RowDefinition Height="30"/>
-            <RowDefinition Height="30"/>
-            <RowDefinition Height="30"/>
-            <RowDefinition Height="*" />
-            <RowDefinition Height="10"/>
-        </Grid.RowDefinitions>
-
-        <!--Top controls-->
-        <TextBlock Text="Configure plugins" FontSize="20" FontWeight="SemiBold" Grid.Column="1" Grid.Row="1" />
-        <TextBlock Text="Select the plugins you want to configure and set the settings for the same" Grid.Column="1" Grid.Row="2" Grid.ColumnSpan="2" />
-
-
-        <!--Left List views-->
-        <ListView Grid.Row="4" Grid.Column="1" Name="lstPluginList" DisplayMemberPath="PluginInstance.PluginName"/>
-
-        <!--Right view-->
-        <Grid Grid.Column="2" Grid.ColumnSpan="2" Grid.Row="3" Grid.RowSpan="5" >
-            <Grid.RowDefinitions>
-                <RowDefinition Height="30" />
-                <RowDefinition Height="*" />
-                <RowDefinition Height="10" />
-            </Grid.RowDefinitions>
-            <CheckBox Name="chkIsEnabled" Grid.Row="0" Grid.Column="2" HorizontalAlignment="Center" Content="IsEnabled" VerticalAlignment="Center" IsChecked="{ Binding ElementName=lstPluginList, Path=SelectedItem.IsEnabled }"/>
-            <ListView Grid.Row="1" ItemsSource="{ Binding ElementName=lstPluginList, Path=SelectedItem.PluginSettings }" IsEnabled="{ Binding ElementName=chkIsEnabled, Path=IsChecked }">
-                <ListView.View>
-                    <GridView>
-                        <GridViewColumn Header="Setting Name" Width="150">
-                            <GridViewColumn.CellTemplate>
-                                <DataTemplate>
-                                    <TextBlock Text="{Binding AttributeData.DisplayName}" TextAlignment="Left" HorizontalAlignment="Left"/>
-                                </DataTemplate>
-                            </GridViewColumn.CellTemplate>
-                        </GridViewColumn>
-                        <GridViewColumn Header="Value" Width="420">
-                            <GridViewColumn.CellTemplate>
-                                <DataTemplate>
-                                    <TextBox Text="{Binding Value}" TextAlignment="Right" HorizontalAlignment="Right" MinWidth="400" TextWrapping="Wrap" AcceptsReturn="True"/>
-                                </DataTemplate>
-                            </GridViewColumn.CellTemplate>
-                        </GridViewColumn>
-                    </GridView>
-                </ListView.View>
-                <ListView.ItemContainerStyle>
-                    <Style TargetType="{x:Type ListViewItem}">
-                        <Setter Property="ToolTip" Value="{Binding AttributeData.HelpText}" />
-                    </Style>
-                </ListView.ItemContainerStyle>
-            </ListView>
-        </Grid>
-    </Grid>
-</Window>
