+++ UI/TunnelRelay.Windows/App.xaml
@@ -1,0 +1,10 @@
+﻿<Application x:Class="TunnelRelay.Windows.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="clr-namespace:TunnelRelay.Windows"
+             StartupUri="Startup.xaml"
+             x:ClassModifier="internal">
+    <Application.Resources>
+         
+    </Application.Resources>
+</Application>
