+++ UI/TunnelRelay.Windows/Engine/HeaderDetails.cs
@@ -1,0 +1,23 @@
+﻿// <copyright file="HeaderDetails.cs" company="Microsoft Corporation">
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+// </copyright>
+
+namespace TunnelRelay.Windows.Engine
+{
+    /// <summary>
+    /// Header details to store header information for UI.
+    /// </summary>
+    public class HeaderDetails
+    {
+        /// <summary>
+        /// Gets or sets the name of the header.
+        /// </summary>
+        public string HeaderName { get; set; }
+
+        /// <summary>
+        /// Gets or sets the header value.
+        /// </summary>
+        public string HeaderValue { get; set; }
+    }
+}
