+++ UI/TunnelRelay.Windows/Startup.xaml
@@ -1,0 +1,12 @@
+﻿<Window x:Class="TunnelRelay.Windows.Startup"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:TunnelRelay.Windows"
+        mc:Ignorable="d"
+        Title="" Height="300" Width="300" ResizeMode="NoResize">
+    <Grid>
+        
+    </Grid>
+</Window>
