+++ UI/TunnelRelay.Windows/UIPages/LoginToAzure.xaml
@@ -1,0 +1,68 @@
+﻿<Window x:Class="TunnelRelay.Windows.LoginToAzure"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:TunnelRelay.Windows"
+        mc:Ignorable="d"
+        Title="" Height="400" Width="700">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="50"/>
+            <RowDefinition Height="60"/>
+            <RowDefinition Height="40"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="15"/>
+            <RowDefinition Height="30"/>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="10"/>
+            <ColumnDefinition Width="*"/>
+            <ColumnDefinition Width="10"/>
+        </Grid.ColumnDefinitions>
+        <TextBlock Text="Tunnel Relay" Grid.Column="1" FontSize="30" FontWeight="SemiBold"/>
+
+        <TextBlock Grid.Row="1" TextWrapping="Wrap" VerticalAlignment="Center" FontSize="20" Grid.Column="1">
+               <Bold>Tunnel Relay</Bold> allows you to expose service running on local machine to the general internet. It uses
+                <Hyperlink 
+                    NavigateUri="https://docs.microsoft.com/en-us/azure/service-bus-relay/relay-what-is-it"
+                    RequestNavigate="Hyperlink_RequestNavigate">
+                    Azure Service Bus
+                </Hyperlink>.
+        </TextBlock>
+
+        <TextBlock Grid.Row="3" Text="You will an Azure subscription to use this application" FontWeight="Bold" FontSize="15" Grid.Column="1"/>
+        <TextBlock Grid.Row="4" FontSize="15" Grid.Column="1">
+            <Hyperlink 
+                    NavigateUri="https://azure.microsoft.com/en-us/free/"
+                    RequestNavigate="Hyperlink_RequestNavigate">
+                    Go here to start an Azure subscription
+                </Hyperlink>.
+        </TextBlock>
+
+        <Button HorizontalAlignment="Right" Content="Login To Azure" Grid.Row="6" Margin="0,0,10,0" Grid.Column="1" Click="LoginToAzure_Click"/>
+
+        <Button HorizontalAlignment="Right" Content="or Import Settings" Grid.Row="7" Margin="0,5,10,0" FontSize="8" Grid.Column="1" Click="BtnImportSettings_Click">
+            <Button.Template>
+                <ControlTemplate TargetType="Button">
+                    <TextBlock TextDecorations="Underline">
+                <ContentPresenter />
+                    </TextBlock>
+                </ControlTemplate>
+            </Button.Template>
+            <Button.Style>
+                <Style TargetType="Button">
+                    <Setter Property="Foreground" Value="Blue" />
+                    <Style.Triggers>
+                        <Trigger Property="IsMouseOver" Value="true">
+                            <Setter Property="Foreground" Value="Red" />
+                        </Trigger>
+                    </Style.Triggers>
+                </Style>
+            </Button.Style>
+        </Button>
+    </Grid>
+</Window>
