+++ UI/TunnelRelay.Windows/UIPages/SelectServiceBus.xaml
@@ -1,0 +1,44 @@
+﻿<Window x:Class="TunnelRelay.Windows.SelectServiceBus"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:TunnelRelay.Windows"
+        mc:Ignorable="d"
+        Title="" Height="400" Width="700" WindowStartupLocation="Manual">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="50"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="30"/>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="10"/>
+            <ColumnDefinition Width="*"/>
+            <ColumnDefinition Width="10"/>
+        </Grid.ColumnDefinitions>
+        <TextBlock Text="Configure" Grid.Column="1" FontSize="25" FontWeight="SemiBold"/>
+        <TextBlock Text="Select the Azure Service Bus" Grid.Column="1" Grid.Row="1" FontSize="20" FontWeight="DemiBold"/>
+        <Label Content="Select the Azure subscription" Grid.Column="1" Grid.Row="2" VerticalAlignment="Bottom" />
+        <ComboBox Name="comboSubscriptionList" Grid.Column="1" Grid.Row="3" DisplayMemberPath="DisplayName" SelectedValuePath="Id" SelectionChanged="SubscriptionList_SelectionChanged" />
+        <Label Content="Select the Azure service bus" Grid.Column="1" Grid.Row="4" VerticalAlignment="Bottom" />
+        <ComboBox Name="comboServiceBusList" Grid.Column="1" Grid.Row="5" DisplayMemberPath="Name" SelectedValuePath="Id" SelectionChanged="ServiceBusList_SelectionChanged" />
+        <Label Name="lblServiceBusName" Content="Enter the name for Service Bus" Grid.Column="1" Grid.Row="6" VerticalAlignment="Bottom" Visibility="Collapsed"/>
+        <TextBox Name="txtServiceBusName" Grid.Column="1" Grid.Row="7" Height="30" Visibility="{Binding ElementName=lblServiceBusName, Path=Visibility }" TextChanged="ServiceBusName_TextChanged"/>
+
+        <StackPanel Orientation="Horizontal" HorizontalAlignment="Right" Grid.Column="1" Grid.Row="9" Name="stcPerformOperation">
+            <ProgressBar Name="progressBar" IsIndeterminate="True" Width="250" Margin="0,0,20,0" Visibility="Hidden"/>
+            <Button Content="Cancel" Width="80" Margin="0,0,20,0" Click="Cancel_Click" />
+            <Button Name="btnDone" Content="Done" Width="80" IsEnabled="False" Click="Done_Click" />
+        </StackPanel>
+    </Grid>
+</Window>
